feat/#35: 저장하기 로직 custom hook으로 분리
